refactor : 스토리북 컨벤션 맞추기
- test 폴더 안에 이름은 해당 컴포넌트 이름으로 모두 변경
