+++ src/components/@common/Modal/CreateGroupModal/test/CreateGroupModal.stories.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { CreateGroupModal } from '.';
+import { CreateGroupModal } from '..';
 
 export default {
   title: 'Component/Modal/CreateGroupModal',
