+++ src/components/@common/Modal/GroupSettingModal/AdminModal/test/AdminModal.stories.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { AdminModal } from '.';
+import { AdminModal } from '..';
 
 export default {
   title: 'Component/Modal/GroupSetting',
