+++ src/components/@common/Modal/OneButtonModal/test/OneButtonModal.stories.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { OneButtonModal } from '.';
+import { OneButtonModal } from '..';
 
 export default {
   title: 'Component/Modal/OneButtonModal',
