+++ src/components/@common/Modal/TwoButtonModal/test/TwoButtonModal.stories.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { TwoButtonModal } from '.';
+import { TwoButtonModal } from '..';
 
 export default {
   title: 'Component/Modal/TwoButtonModal',
